refactoring continues: naming(namespaces), logic separation
