+++ lib/modules/validation.rb
@@ -1,0 +1,7 @@
+module Validation
+  def validate_input(input)
+    validation = self.class::Schema.call(input)
+
+    validation.success? ? Success(input) : Failure(validation.messages)
+  end
+end
