+++ lib/readers/file_reader_error.rb
@@ -1,0 +1,7 @@
+module Readers
+  class FileReaderError < StandardError
+    def initialize(msg)
+      super(msg)
+    end
+  end
+end
