--- spec/parsers/affix_parser_spec.rb
@@ -1,4 +1,0 @@
-require_relative '../../lib/parsers/affix_parser'
-
-RSpec.describe Parsers::AffixParser do
-end
