--- spec/parsers/dictionary_reader_args_parser_spec.rb
@@ -1,4 +1,0 @@
-require_relative '../../lib/parsers/dictionary_reader_args_parser'
-
-RSpec.describe Parsers::DictionaryReaderArgsParser do
-end
