+++ spec/readers/file_reader_spec.rb
@@ -1,0 +1,2 @@
+RSpec.describe Readers::FileReader do
+end
